backend/requirements.txt
fastapi
uvicorn[standard]
whisper-openai
torch
torchaudio
pydantic
python-multipart
